Fix three real defects surfaced by clang 21, and clear all 407 warnings
Apple clang 21 is stricter than what this tree was last built with. Three of
the diagnostics were genuine bugs:

primitiveObjectPointsTo (prim 132) guarded the word-array case with
`format >= Indexable32 && format <= Indexable64`, i.e. `>= 10 && <= 9`, which
is never true. WordArray/DoubleWordArray/Bitmap/Form bits fell through to the
pointer-slot loop, which reinterprets raw 32/64-bit pixel data as oops and
compares it against the target — so `anObject pointsTo:` could answer true
purely because a bitmap contained the target oop's bit pattern. Bounds
corrected to 9..11. (The 16-bit formats still have the identical defect; left
alone pending a decision, noted in docs/changes.md.)

MIDIPlugin memset over OpenPort, which embeds a std::mutex touched from the
CoreMIDI input callback thread. Zeroing a live pthread mutex is UB. Now a
resetPort() helper clears only the POD fields.

PlatformBridge wrote #pragma push_macro("nil") / pop_macro("nil") as though
they were scoped to control flow, one pop before each early return. They are
preprocessor-lexical, so the first pop re-enabled ObjC's nil macro for the
remaining ~28 lines meant to run without it. Now one push, one pop.

For the rest: 315 of the 407 warnings were in B2DPlugin.c, JPEGReaderPlugin.c
and DSAPrims.c, which VMMaker generates from OpenSmalltalk-VM sources — edits
there would be lost on regeneration, so they now build with -w, matching how
bundled libjpeg and testlib were already handled. The remaining 92 were in
hand-written code and were fixed, not suppressed: dead locals removed after
checking their initializers had no side effects, mis-ordered ctor init lists
reordered, wrong printf specifiers corrected, and the meaningless `volatile`
dropped from utcMicroseconds' return type in sqVirtualMachine.h (with the
matching change in InterpreterProxy.cpp — in C++ the cv-qualifier survives in
the function-pointer type, so both halves must move together).

PharoVMCore now builds warning-free. Note that a full build still needs
Frameworks/ populated by scripts/build-third-party.sh; that directory is
gitignored and absent here, so this was verified compile-only against
Homebrew's libffi and SDL2 headers.

Several things found along the way are reported but NOT fixed, because they
change GC semantics and warrant a decision: ObjectMemory::fullGC ignores
planCompactSavingForwarders() returning false on eden overflow and compacts
on a partial plan anyway; sweepGC's fullSize adds 8 bytes that
ObjectHeader::totalSize() already includes; and GCResult::objectsMoved is
never populated, so every GC report claims "moved 0 objects".

+++ src/vm/Interpreter.hpp
@@ -377,7 +377,10 @@ public:
     bool primitiveSucceeded() const { return !primitiveFailed_; }
 
     // External plugin primitive support (for B2DPlugin, etc.)
-    using ExternalPrimFunc = int (*)(void);
+    // Return type must be sqInt (`long`, see plugins/sqMemoryAccess.h): that is what
+    // every VMMaker-generated plugin entry point returns. Declaring it as `int` here
+    // made every registration site cast between incompatible function types.
+    using ExternalPrimFunc = long (*)(void);
     void registerNamedPrimitive(const std::string& module, const std::string& name, ExternalPrimFunc func);
     PrimitiveResult callExternalPrimitive(ExternalPrimFunc fn);
